Adiciona factory de product a seeder

+++ database/migrations/2023_11_08_131943_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nome');
             $table->string('descricao');
-            $table->float("preco", 2);
+            $table->double("preco");
             $table->integer("quantidade");
             $table->timestamps();
         });
